Add insert shoe functionality
AddShoeProcess controller code done. Shoe model modified to work with
our insert, specifically the id paramater was not passed through as it
is auto-incremented and the save function was modified to reflect that.
The Update is to be put into it's own function and the database variable
needs to changed to call the getPDO function in Rapid.