Fix inputs for adding and updating a shoe
This commit adds the ability to go up in steps for the size and price of
a shoe within the form. It also changes the values for updating a shoe
to display the physical values instead of a placeholder so the user
doesn't have to re-enter data that they wanted to keep.