Add a session for a users role upon login
This commit adds a way for us to check a users role when they are logged
in so that we can direct different content on the website towards
different users based on their role. For e.g. A normal user with the
role 'user' will not see any administrative features that a manager with
the role of 'manager' will see.