Add url to Shoe and changed products
A url column was added to the database to allow linking of shoes to
images in the products page.The Shoe model had url added to it to
accommodate the new column. Need to fix the display next and change the
add, update and display shoes to accommodate the new column.